GeminiTuneOpsApp/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
google-auth==2.35.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-bigquery-storage==2.27.0
google-cloud-storage==2.18.2
pandas==2.2.3
pyarrow==18.0.0
scikit-learn==1.5.2
